+++ banking-api/src/main/java/banking_api/service/TransferService.java
@@ -1,13 +1,19 @@
 package banking_api.service;
 
+
 import banking_api.dto.TransactionResponse;
 import banking_api.model.User;
 
+import java.math.BigDecimal;
+
+
 public interface TransferService {
+
 
     TransactionResponse transfer(
             User sender,
             String receiverAccountNumber,
-            Double amount
+            BigDecimal amount
     );
+
 }
